Move most of signal handling to Ryujinx.Cpu project
