+++ src/Ryujinx.Cpu/Jit/JitMemoryAllocator.cs
@@ -7,7 +7,5 @@ namespace Ryujinx.Cpu.Jit
     {
         public IJitMemoryBlock Allocate(ulong size) => new JitMemoryBlock(size, MemoryAllocationFlags.None);
         public IJitMemoryBlock Reserve(ulong size) => new JitMemoryBlock(size, MemoryAllocationFlags.Reserve | MemoryAllocationFlags.Jit);
-
-        public ulong GetPageSize() => MemoryBlock.GetPageSize();
     }
 }
